See description
The variable "star_specific_flux" was chaged to "star_specific_photons_er_second".

There are no ordinary and extraordinary rays anymore. These two variables were added together in one unique variable.

Now, the AIS takes into account the transmission curves of almost all optical components. Also, the AIS applies the transformation matrices of the polarimetric components.

For the retardes, the AIS calculates the matrix that corresponde to the phase difference as a function of the wavelength